+++ bin/miden-cli/src/commands/init.rs
@@ -1,7 +1,13 @@
-use std::{fs::File, io::Write, path::PathBuf, str::FromStr};
+use std::{
+    fs::{self, File},
+    io::Write,
+    path::PathBuf,
+    str::FromStr,
+};
 
 use clap::Parser;
 use miden_client::rpc::Endpoint;
+use tracing::info;
 
 use crate::{
     config::{CliConfig, CliEndpoint},
@@ -9,7 +15,17 @@ use crate::{
     CLIENT_CONFIG_FILE_NAME,
 };
 
-// Init COMMAND
+/// Contains the account component template file generated on build.rs, corresponding to the
+/// fungible faucet component.
+const FAUCET_TEMPLATE_FILE: &[u8] =
+    include_bytes!(concat!(env!("OUT_DIR"), "/templates/basic-fungible-faucet.mct"));
+
+/// Contains the account component template file generated on build.rs, corresponding to the basic
+/// auth component.
+const BASIC_AUTH_TEMPLATE_FILE: &[u8] =
+    include_bytes!(concat!(env!("OUT_DIR"), "/templates/basic-auth.mct"));
+
+// INIT COMMAND
 // ================================================================================================
 
 #[derive(Debug, Clone)]
@@ -49,7 +65,7 @@ impl Network {
 #[clap(
     about = "Initialize the client. It will create a file named `miden-client.toml` that holds \
 the CLI and client configurations, and will be placed by default in the current working \
-directory."
+directory"
 )]
 pub struct InitCmd {
     /// Network configuration to use. Options are `devnet`, `testnet`, `localhost` or a custom RPC
@@ -112,6 +128,8 @@ impl InitCmd {
             CliError::Config("failed to create config file".to_string().into(), err.to_string())
         })?;
 
+        write_template_files(&cli_config)?;
+
         file_handle.write(config_as_toml_string.as_bytes()).map_err(|err| {
             CliError::Config("failed to write config file".to_string().into(), err.to_string())
         })?;
@@ -120,4 +138,33 @@ impl InitCmd {
 
         Ok(())
     }
+}
+
+/// Creates the directory specified by `cli_config.component_template_directory`
+/// and writes the default included component templates.
+fn write_template_files(cli_config: &CliConfig) -> Result<(), CliError> {
+    fs::create_dir_all(&cli_config.component_template_directory).map_err(|err| {
+        CliError::Config(
+            "failed to create account component templates directory".to_string().into(),
+            err.to_string(),
+        )
+    })?;
+
+    // Write the faucet template file.
+    // TODO: io errors should probably have their own context.
+    let faucet_template_path =
+        cli_config.component_template_directory.join("basic-fungible-faucet.mct");
+    let mut faucet_file = File::create(&faucet_template_path)?;
+    faucet_file.write_all(FAUCET_TEMPLATE_FILE)?;
+
+    let basic_auth_template_path = cli_config.component_template_directory.join("basic-auth.mct");
+    let mut basic_auth_file = File::create(&basic_auth_template_path)?;
+    basic_auth_file.write_all(BASIC_AUTH_TEMPLATE_FILE)?;
+
+    info!(
+        "Template files successfully created in: {:?}",
+        cli_config.component_template_directory
+    );
+
+    Ok(())
 }